tests/gtrsim.dat

#data
&gtrsim
#errors
#document
| <html>
|   <head>
|   <body>
|     ">rsim"
